+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to
 - ♿(frontend) improve accessibility:
   - #1354
   - ♿ improve accessibility by adding landmark roles to layout #1394
-  - ✨ add document visible in list and openable via enter key #1365
+  - ♿ add document visible in list and openable via enter key #1365
   - ♿ add pdf outline property to enable bookmarks display #1368
 
 ### Fixed
@@ -28,6 +28,8 @@ and this project adheres to
 ### Added
 
 - ✨(api) add API route to fetch document content #1206
+- ♿(frontend) improve accessibility:
+  - #1349
 
 ### Changed